backend/requirements.txt
# Manga Colorizer - Dependencies

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# For manga-colorization-v2 (Fast engine)
gdown>=4.7.0
scikit-image>=0.21.0

# Batch processing
tqdm>=4.65.0
natsort>=8.4.0

# Manga downloading
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# Browser automation for MangaFire VRF token extraction
playwright>=1.40.0
nest-asyncio>=1.5.8

# FastAPI backend for Electron app
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
websockets>=12.0

# Note: On Mac, install PyTorch with MPS support:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
